Added Fetch data from the Mock DB
