+++ FrontEnd/db.json
@@ -1,20 +1,103 @@
 {
   "eventsList": [
     {
-      "id": "1",
+      "image_id": "assets/Image1.jpg",
       "title": "ABC Party",
-      "img": "assets/Image1.jpg",
-      "hostname": "Ravi",
-      "distance": "0.6 miles away",
-      "count": "20"
+      "host_name": "Ravi",
+      "distance_miles": 0.6,
+      "interested_people": 20,
+      "alcohol": true,
+      "over_18": true
     },
     {
-      "id": "2",
-      "title": "HouseParty",
-      "img": "assets/Image2.jpg",
-      "hostname": "John",
-      "distance": "0.2 miles away",
-      "count": "50"
+      "image_id": "assets/Image2.jpg",
+      "title": "ABC Party",
+      "host_name": "Ravi",
+      "distance_miles": 0.6,
+      "interested_people": 20,
+      "alcohol": true,
+      "over_18": true
+    },
+    {
+      "image_id": "assets/Image3.jpg",
+      "title": "ABC Party",
+      "host_name": "Ravi",
+      "distance_miles": 0.6,
+      "interested_people": 20,
+      "alcohol": true,
+      "over_18": true
+    },
+    {
+      "image_id": "assets/Image4.jpg",
+      "title": "ABC Party",
+      "host_name": "Ravi",
+      "distance_miles": 0.6,
+      "interested_people": 20,
+      "alcohol": true,
+      "over_18": true
+    },
+    {
+      "image_id": "assets/Image5.jpg",
+      "title": "ABC Party",
+      "host_name": "Ravi",
+      "distance_miles": 0.6,
+      "interested_people": 20,
+      "alcohol": true,
+      "over_18": true
+    },
+    {
+      "image_id": "assets/Image6.jpg",
+      "title": "ABC Party",
+      "host_name": "Ravi",
+      "distance_miles": 0.6,
+      "interested_people": 20,
+      "alcohol": true,
+      "over_18": true
+    },
+    {
+      "image_id": "assets/Image7.jpg",
+      "title": "ABC Party",
+      "host_name": "Ravi",
+      "distance_miles": 0.6,
+      "interested_people": 20,
+      "alcohol": true,
+      "over_18": true
+    },
+    {
+      "image_id": "assets/Image8.jpg",
+      "title": "ABC Party",
+      "host_name": "Ravi",
+      "distance_miles": 0.6,
+      "interested_people": 20,
+      "alcohol": true,
+      "over_18": true
+    },
+    {
+      "image_id": "assets/Image9.jpg",
+      "title": "ABC Party",
+      "host_name": "Ravi",
+      "distance_miles": 0.6,
+      "interested_people": 20,
+      "alcohol": true,
+      "over_18": true
+    },
+    {
+      "image_id": "assets/Image10.jpg",
+      "title": "ABC Party",
+      "host_name": "Ravi",
+      "distance_miles": 0.6,
+      "interested_people": 20,
+      "alcohol": true,
+      "over_18": true
+    },
+    {
+      "image_id": "assets/Image11.jpg",
+      "title": "ABC Party",
+      "host_name": "Ravi",
+      "distance_miles": 0.6,
+      "interested_people": 20,
+      "alcohol": true,
+      "over_18": true
     }
   ]
 }
